refactor: moved type to appropriate folder
